Revert "Merge branch 'main' of https://github.com/CodeLikeCXK/DagorEngine"
This reverts commit 416b43f4c108c0e4c425e782e341e3021eca9c9a, reversing
changes made to 982264f2a1390b6f841f9fc4845acdae57ca909c.

+++ prog/tools/libTools/dtx/dtxSampler.cpp
@@ -1,9 +1,7 @@
-// Copyright (C) Gaijin Games KFT.  All rights reserved.
-
 #include <libTools/dtx/dtxSampler.h>
 // #include <debug/dag_debug.h>
 #include "rgbe.h"
-#include <drv/3d/dag_tex3d.h>
+#include <3d/dag_tex3d.h>
 #include <3d/ddsFormat.h>
 
 
@@ -78,7 +76,7 @@ static void mask_to_ofs(int mask, int &ofs, int &shift, int target_sz)
     shift = 0;
     return;
   }
-  unsigned long o, s;
+  DWORD o, s;
 #if _TARGET_64BIT
   o = s = 32;
   __bit_scan_forward(o, mask);
